+++ Demo/src/main/java/com/paypal/android/ui/paypalbuttons/PayPalButtonsViewModel.kt
@@ -46,7 +46,12 @@ class PayPalButtonsViewModel : ViewModel() {
     var paymentButtonShape: PaymentButtonShape
         get() = _uiState.value.paymentButtonShape
         set(value) {
-            _uiState.update { it.copy(paymentButtonShape = value) }
+            _uiState.update {
+                it.copy(
+                    paymentButtonShape = value,
+                    customCornerRadius = null
+                )
+            }
         }
 
     var paymentButtonSize: PaymentButtonSize